refactor(contacts): move add contact drawer to flow
