+++ features/flow/contacts/src/add/index.ts
@@ -2,3 +2,10 @@ export type { ContactCreationInput, ContactCreationPort } from "./ports";
 export type { AddContactViewModel } from "./types";
 export { createAddContactViewModel } from "./viewModel";
 export { createAddContactController, type AddContactController } from "./controller";
+export { useAddContactDrawerViewModel } from "./useAddContactDrawerViewModel";
+export type {
+  AddContactDrawerViewModel,
+  ContactsAddContactDrawerLabels,
+  ContactsAddContactDrawerProps,
+  UseAddContactDrawerViewModelOptions,
+} from "./drawer.types";
