+++ features/flow/contacts/src/add/native.ts
@@ -1,0 +1,1 @@
+export { ContactsAddContactDrawer } from "./components/ContactsAddContactDrawer.native";
