+++ features/flow/contacts/src/jest.native.ts
@@ -1,5 +1,6 @@
 export * from "./add";
 export * from "./featureIntroduction";
+export * from "./add/native";
 export * from "./featureFlags";
 export * from "./hooks";
 export * from "./list";
